Rename selectMenu to KnightHacksRolesMenu
