+++ src/common/KnightHacksRolesMenu.ts
